ADD method to edit a product

+++ SimpleInventorySystem/Product.cs
@@ -4,9 +4,9 @@ namespace SimpleInventorySystem
     {
         public string Name { get; set; }
         public int Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public Product(string name, int quantity, double price)
+        public Product(string name, int quantity, decimal price)
         {
             Name = name;
             Quantity = quantity;
